backend/requirements.txt
Django==2.2
django-cors-headers==3.2.1
django-extensions==2.2.6
djangorestframework==3.11.0
gunicorn==20.0.4
numpy==1.19.0
pandas==1.0.5
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.0
whitenoise==5.1.0
xlrd==1.2.0
XlsxWriter==1.2.9
